Factor out lambda-response-http-headers.tsx.
